test: move format_code tests to new file
