+++ test_docformatter.py
@@ -45,663 +45,6 @@ else:
                                 'docformatter.py')]  # pragma: no cover
 
 
-class TestUnits(unittest.TestCase):
-
-    def test_format_code(self):
-        self.assertEqual(
-            '''\
-def foo():
-    """Hello foo."""
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-    """
-    Hello foo.
-    """
-'''))
-
-    def test_format_code_range_miss(self):
-        self.assertEqual('''\
-def f(x):
-    """  This is a docstring. That should be on more lines"""
-    pass
-def g(x):
-    """  Badly indented docstring"""
-    pass''',
-                         docformatter.format_code('''\
-def f(x):
-    """  This is a docstring. That should be on more lines"""
-    pass
-def g(x):
-    """  Badly indented docstring"""
-    pass''', line_range=[1, 1]))
-
-    def test_format_code_range_hit(self):
-        self.assertEqual('''\
-def f(x):
-    """This is a docstring.
-
-    That should be on more lines
-    """
-    pass
-def g(x):
-    """  Badly indented docstring"""
-    pass''',
-                         docformatter.format_code('''\
-def f(x):
-    """  This is a docstring. That should be on more lines"""
-    pass
-def g(x):
-    """  Badly indented docstring"""
-    pass''', line_range=[1, 2]))
-
-    def test_format_code_docstring_length(self):
-        self.assertEqual('''\
-def f(x):
-    """This is a docstring.
-
-
-    That should be on less lines
-    """
-    pass
-def g(x):
-    """Badly indented docstring."""
-    pass''',
-                         docformatter.format_code('''\
-def f(x):
-    """This is a docstring.
-
-
-    That should be on less lines
-    """
-    pass
-def g(x):
-    """  Badly indented docstring"""
-    pass''', length_range=[1, 1]))
-
-    def test_format_code_with_module_docstring(self):
-        self.assertEqual(
-            '''\
-#!/usr/env/bin python
-"""This is a module docstring.
-
-1. One
-2. Two
-"""
-
-"""But
-this
-is
-not."""
-''',
-            docformatter.format_code(
-                '''\
-#!/usr/env/bin python
-"""This is
-a module
-docstring.
-
-1. One
-2. Two
-"""
-
-"""But
-this
-is
-not."""
-'''))
-
-    def test_format_code_should_ignore_non_docstring(self):
-        source = '''\
-x = """This
-is
-not."""
-'''
-        self.assertEqual(
-            source,
-            docformatter.format_code(source))
-
-    def test_format_code_with_empty_string(self):
-        self.assertEqual(
-            '',
-            docformatter.format_code(''))
-
-    def test_format_code_with_tabs(self):
-        self.assertEqual(
-            '''\
-def foo():
-\t"""Hello foo."""
-\tif True:
-\t\tx = 1
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-\t"""
-\tHello foo.
-\t"""
-\tif True:
-\t\tx = 1
-'''))
-
-    def test_format_code_with_mixed_tabs(self):
-        self.assertEqual(
-            '''\
-def foo():
-\t"""Hello foo."""
-\tif True:
-\t    x = 1
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-\t"""
-\tHello foo.
-\t"""
-\tif True:
-\t    x = 1
-'''))
-
-    def test_format_code_with_escaped_newlines(self):
-        self.assertEqual(
-            r'''def foo():
-    """Hello foo."""
-    x = \
-            1
-''',
-            docformatter.format_code(
-                r'''def foo():
-    """
-    Hello foo.
-    """
-    x = \
-            1
-'''))
-
-    def test_format_code_with_comments(self):
-        self.assertEqual(
-            r'''
-def foo():
-    """Hello foo."""
-    # My comment
-    # My comment with escape \
-    123
-'''.lstrip(),
-            docformatter.format_code(
-                r'''
-def foo():
-    """
-    Hello foo.
-    """
-    # My comment
-    # My comment with escape \
-    123
-'''.lstrip()))
-
-    def test_format_code_with_escaped_newline_in_inline_comment(self):
-        self.assertEqual(
-            r'''
-def foo():
-    """Hello foo."""
-    def test_method_no_chr_92(): the501(92) # \
-'''.lstrip(),
-            docformatter.format_code(
-                r'''
-def foo():
-    """
-    Hello foo.
-    """
-    def test_method_no_chr_92(): the501(92) # \
-'''.lstrip()))
-
-    def test_format_code_raw_docstring_double_quotes(self):
-        """Should format raw docstrings with triple double quotes.
-
-        See requirement #2.  See issue #54 for request to handle raw
-        docstrings.
-        """
-        self.assertEqual(
-            '''\
-def foo():
-    r"""Hello foo."""
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-    r"""
-    Hello foo.
-    """
-'''))
-
-        self.assertEqual(
-        '''\
-def foo():
-    R"""Hello foo."""
-''',
-            docformatter.format_code(
-            '''\
-def foo():
-    R"""
-    Hello foo.
-    """
-'''))
-
-    def test_format_code_raw_docstring_single_quotes(self):
-        """Should format raw docstrings with triple single quotes.
-
-        See requirement #2.  See issue #54 for request to handle raw
-        docstrings.
-        """
-        self.assertEqual(
-            '''\
-def foo():
-    r"""Hello foo."""
-''',
-            docformatter.format_code(
-                """\
-def foo():
-    r'''
-    Hello foo.
-    '''
-"""))
-
-        self.assertEqual(
-        '''\
-def foo():
-    R"""Hello foo."""
-''',
-           docformatter.format_code(
-            """\
-def foo():
-    R'''
-    Hello foo.
-    '''
-"""))
-
-    def test_format_code_unicode_docstring_double_quotes(self):
-        """Should format unicode docstrings with triple double quotes.
-
-        See requirement #3.  See issue #54 for request to handle raw
-        docstrings.
-        """
-        self.assertEqual(
-            '''\
-def foo():
-    u"""Hello foo."""
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-    u"""
-    Hello foo.
-    """
-'''))
-
-        self.assertEqual(
-            '''\
-def foo():
-    U"""Hello foo."""
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-    U"""
-    Hello foo.
-    """
-'''))
-
-    def test_format_code_unicode_docstring_single_quotes(self):
-        """Should format unicode docstrings with triple single quotes.
-
-        See requirement #3.  See issue #54 for request to handle raw
-        docstrings.
-        """
-        self.assertEqual(
-            '''\
-def foo():
-    u"""Hello foo."""
-''',
-            docformatter.format_code(
-                """\
-def foo():
-    u'''
-    Hello foo.
-    '''
-"""))
-
-        self.assertEqual(
-            '''\
-def foo():
-    U"""Hello foo."""
-''',
-            docformatter.format_code(
-                """\
-def foo():
-    U'''
-    Hello foo.
-    '''
-"""))
-
-    def test_format_code_skip_nested(self):
-        code = """\
-def foo():
-    '''Hello foo. \"\"\"abc\"\"\"
-    '''
-"""
-        self.assertEqual(code, docformatter.format_code(code))
-
-    def test_format_code_with_multiple_sentences(self):
-        self.assertEqual(
-            '''\
-def foo():
-    """Hello foo.
-
-    This is a docstring.
-    """
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-    """
-    Hello foo.
-    This is a docstring.
-    """
-'''))
-
-    def test_format_code_with_multiple_sentences_same_line(self):
-        self.assertEqual(
-            '''\
-def foo():
-    """Hello foo.
-
-    This is a docstring.
-    """
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-    """
-    Hello foo. This is a docstring.
-    """
-'''))
-
-    def test_format_code_with_multiple_sentences_multi_line_summary(self):
-        self.assertEqual(
-            '''\
-def foo():
-    """Hello foo.
-
-    This is a docstring.
-    """
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-    """
-    Hello
-    foo. This is a docstring.
-    """
-'''))
-
-    def test_format_code_with_empty_lines(self):
-        self.assertEqual(
-            '''\
-def foo():
-    """Hello foo and this is a docstring.
-
-    More stuff.
-    """
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-    """
-    Hello
-    foo and this is a docstring.
-
-    More stuff.
-    """
-'''))
-
-    def test_format_code_with_trailing_whitespace(self):
-        self.assertEqual(
-            '''\
-def foo():
-    """Hello foo and this is a docstring.
-
-    More stuff.
-    """
-''',
-            docformatter.format_code(
-                ('''\
-def foo():
-    """
-    Hello
-    foo and this is a docstring.\t
-
-    More stuff.\t
-    """
-''')))
-
-    def test_format_code_with_parameters_list(self):
-        self.assertEqual(
-            '''\
-def foo():
-    """Test.
-
-    one - first
-    two - second
-    """
-''',
-            docformatter.format_code(
-                ('''\
-def foo():
-    """Test
-    one - first
-    two - second
-    """
-''')))
-
-    def test_ignore_code_with_single_quote(self):
-        """Single single quote on first line of code should remain untouched.
-
-        See requirement 1, always use triple double quotes.
-        See issue #66 for example of docformatter breaking code when encountering
-        single quote.
-        """
-        self.assertEqual(
-                '''\
-def foo():
-    'Just a regular string'
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-    'Just a regular string'
-'''))
-
-    def test_ignore_code_with_double_quote(self):
-        """Single double quotes on first line of code should remain untouched.
-
-        See requirement 1, always use triple double quotes.
-        See issue #66 for example of docformatter breaking code when encountering
-        single quote.
-        """
-        self.assertEqual(
-                '''\
-def foo():
-    "Just a regular string"
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-    "Just a regular string"
-'''))
-
-    def test_format_code_with_should_skip_nested_triple_quotes(self):
-        line = '''\
-def foo():
-    'Just a """foo""" string'
-'''
-        self.assertEqual(line, docformatter.format_code(line))
-
-    def test_format_code_with_assignment_on_first_line(self):
-        self.assertEqual(
-            '''\
-def foo():
-    x = """Just a regular string. Alpha."""
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-    x = """Just a regular string. Alpha."""
-'''))
-
-    def test_format_code_with_regular_strings_too(self):
-        self.assertEqual(
-            '''\
-def foo():
-    """Hello foo and this is a docstring.
-
-    More stuff.
-    """
-    x = """My non-docstring
-    This should not touched."""
-
-    """More stuff
-    that should not be
-    touched\t"""
-''',
-            docformatter.format_code(
-                '''\
-def foo():
-    """
-    Hello
-    foo and this is a docstring.
-
-    More stuff.
-    """
-    x = """My non-docstring
-    This should not touched."""
-
-    """More stuff
-    that should not be
-    touched\t"""
-'''))
-
-    def test_format_code_with_syntax_error(self):
-        self.assertEqual('"""\n',
-                         docformatter.format_code('"""\n'))
-
-    def test_format_code_with_syntax_error_case_slash_r(self):
-        self.assertEqual('"""\r',
-                         docformatter.format_code('"""\r'))
-
-    def test_format_code_with_syntax_error_case_slash_r_slash_n(self):
-        self.assertEqual('"""\r\n',
-                         docformatter.format_code('"""\r\n'))
-
-    def test_format_code_dominant_line_ending_style_preserved(self):
-        goes_in = '''\
-def foo():\r
-    """\r
-    Hello\r
-    foo. This is a docstring.\r
-    """\r
-'''
-        self.assertEqual(docformatter.CRLF, docformatter.find_newline(goes_in.splitlines(True)))
-        self.assertEqual(
-            '''\
-def foo():\r
-    """Hello foo.\r
-\r
-    This is a docstring.\r
-    """\r
-''',
-            docformatter.format_code(goes_in))
-
-    def test_format_code_additional_empty_line_before_doc(self):
-        args = {'summary_wrap_length': 79,
-                'description_wrap_length': 72,
-                'pre_summary_newline': False,
-                'make_summary_multi_line': False,
-                'post_description_blank': False,
-                'force_wrap': False,
-                'line_range': None}
-        self.assertEqual('\n\n\ndef my_func():\n"""Summary of my function."""\npass',
-                         docformatter._format_code('\n\n\ndef my_func():\n\n"""Summary of my function."""\npass', args))
-
-    def test_format_code_extra_newline_following_comment(self):
-        docstring = ('''\
-def crash_rocket(location):    # pragma: no cover
-
-    """This is a docstring following an in-line comment."""
-    return location''')
-        self.assertEqual('''\
-def crash_rocket(location):    # pragma: no cover
-    """This is a docstring following an in-line comment."""
-    return location''',
-                         docformatter._format_code(docstring))
-
-    def test_format_code_no_docstring(self):
-        args = {'summary_wrap_length': 79,
-                'description_wrap_length': 72,
-                'pre_summary_newline': False,
-                'pre_summary_space': False,
-                'make_summary_multi_line': False,
-                'post_description_blank': False,
-                'force_wrap': False,
-                'line_range': None}
-        docstring = ("def pytest_addoption(parser: pytest.Parser) -> "
-        "None:\n    register_toggle.pytest_addoption(parser)\n")
-        self.assertEqual(docstring,
-                         docformatter._format_code(docstring, args))
-
-        docstring = ("def pytest_addoption(parser: pytest.Parser) -> "
-                     "None:    # pragma: no cover\n    "
-                     "register_toggle.pytest_addoption(parser)\n")
-        self.assertEqual(docstring,
-                         docformatter._format_code(docstring, args))
-
-    def test_exclude(self):
-        sources = {"/root"}
-        patch_data = [
-            ("/root", ['folder_one', 'folder_two'], []),
-            ("/root/folder_one", ['folder_three'], ["one.py"]),
-            ("/root/folder_one/folder_three", [], ["three.py"]),
-            ("/root/folder_two", [], ["two.py"]),
-        ]
-        with patch("os.walk", return_value=patch_data), patch("os.path.isdir", return_value=True):
-            test_exclude_one = list(docformatter.find_py_files(sources, True, ["folder_one"]))
-            self.assertEqual(test_exclude_one, ['/root/folder_two/two.py'])
-            test_exclude_two = list(docformatter.find_py_files(sources, True, ["folder_two"]))
-            self.assertEqual(test_exclude_two, ['/root/folder_one/one.py', '/root/folder_one/folder_three/three.py'])
-            test_exclude_three = list(docformatter.find_py_files(sources, True, ["folder_three"]))
-            self.assertEqual(test_exclude_three, ['/root/folder_one/one.py', '/root/folder_two/two.py'])
-            test_exclude_py = list(docformatter.find_py_files(sources, True, ".py"))
-            self.assertFalse(test_exclude_py)
-            test_exclude_two_and_three = list(docformatter.find_py_files(sources, True, ["folder_two", "folder_three"]))
-            self.assertEqual(test_exclude_two_and_three, ['/root/folder_one/one.py'])
-            test_exclude_files = list(docformatter.find_py_files(sources, True, ["one.py", "two.py"]))
-            self.assertEqual(test_exclude_files, ['/root/folder_one/folder_three/three.py'])
-
-    def test_exclude_nothing(self):
-        sources = {"/root"}
-        patch_data = [
-            ("/root", ['folder_one', 'folder_two'], []),
-            ("/root/folder_one", ['folder_three'], ["one.py"]),
-            ("/root/folder_one/folder_three", [], ["three.py"]),
-            ("/root/folder_two", [], ["two.py"]),
-        ]
-        with patch("os.walk", return_value=patch_data), patch("os.path.isdir", return_value=True):
-            test_exclude_nothing = list(docformatter.find_py_files(sources, True, []))
-            self.assertEqual(test_exclude_nothing, ['/root/folder_one/one.py', '/root/folder_one/folder_three/three.py',
-                                                    '/root/folder_two/two.py'])
-            test_exclude_nothing = list(docformatter.find_py_files(sources, True))
-            self.assertEqual(test_exclude_nothing, ['/root/folder_one/one.py', '/root/folder_one/folder_three/three.py',
-                                                    '/root/folder_two/two.py'])
-
 class TestSystem(unittest.TestCase):
 
     def test_diff(self):
